linux-firmware-wl12xx: install the latest wl12xx mac80211 firmware files
The mac80211 firmware files that are currently fetched by the
linux-firmware_git recipe in open-embedded are not yet updated to the
ones that should be used with our latest NLCP release.

Due to this and as a temporary solution we need to fetch the latest
firmware files from our GForge page.

This is only a temporary solution until the latest firmware files will be
in the linux-firmware git.
We will remove this amend once the latest firmware files are updated in
the git.

Signed-off-by: Eyal Reizer <eyalr@ti.com>
Acked-by: Chase Maupin <chase.maupin@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>