wpa-gui-e: fix network configuration dialog
The network config dialog in the wpa gui does not fit the screen size
used on the am37x-evm and the am180x-evm boards, and is therfore unusable.
This patch makes the network config dialog scrollable when maximized and
allows controlling it properly.

Upstream-Status: Pending

We need this patch now in Arago in order to be able to use the wpa-gui-e in the
coming am37x-evm, am180x-evm SDK release.

Signed-off-by: Eyal Reizer <eyalr@ti.com>
Acked-by: Chase Maupin <chase.maupin@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>