linux: remove kernel AUTOREV'ed recipes
AUTOREV'ed recipes were causing many build failures even if the
recipes are not used in the build process because bitbake will
parse all the recipes before it starts building and AUTOREV'ed
recipes will try to list the repository status to get the SRCREV
during the parse step. If the url used by one of the AUTOREV
recipes is down, the parsing step will fail and nothing will get
built.
The AUTOREV kernel recipes will move to a test layer.

Signed-off-by: Carlos Hernandez <c-hernandez8@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>