u-boot: updated u-boot 2011.09 branch
* Updated to the latest SDK integration branch and SRCREV.
* Bumped the PR

Notes from Tom:
a) The prompt is now 'U-Boot# '.  Update any scripts to match
b) As of yet full Beaglebone support is not merged so for example eth
will not work there yet (the goal is very soon however).
c) If you have NOT used SPL before, we now use u-boot.img not
u-boot.bin.  Again, u-boot.img is the file to use.

Please note:
0) You can now build with O= (or BUILD_DIR=) which will put generated
files in that subdir and keep your tree clean.  This is good.
1) With the switch to SPL, 'MLO' comes out of the build, usable, at the end.
2) The SPL binary we have will then read u-boot.img from NAND or MMC,
depending on what the romcode tells us SPL was loaded from.
3) For more general information about SPL, see docs/README.SPL in the repo.
4) We have uEnv.txt support, in the same manner as beagleboard, without
'user button' support, but uenvcmd works of course.
5) If you previously had a "No NAND" board, that may be fixed.  There's
still reports of some boards not having NAND with this but there's a
number of "now it works" reports.

Signed-off-by: Chase Maupin <Chase.Maupin@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>