libgles-omap3: Setting of WindowSystem back to FRONT mode
* Bumped up the PR
* There are two problems with setting it to QWS WSEGL:
1. Earlier SDK releases had this set to FRONT mode. The Qt's QWS will default
to FLIP mode. This will create an impression that graphics SDKs have gone back
in performance from one mode to another.

2. Qt's QWS WSEGL has issues with 3D demos on Qt 4.7. When the SGX demos quit,
it results in a glibc double free issue. There are no problems with Qt being
accelerated with SGX. This issue has been newly introduced in Qt 4.7.2 and was
not observed with Qt 4.6.3

Signed-off-by: Anand Balagopalakrishnan <anandb@ti.com>
Acked-by: Chase Maupin <chase.maupin@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>